Fixed bug in point2mesh's support for applying a rotation.
Added --flip-triangles option to point2mesh for .iv and .tri output
which saves the mesh using clockwise vertex ordering.

Updated StereoSessionKeypoint.cc to be compatible with changes in the
InterestPoint and Math interfaces in VW. The code is essentially the
same as in ipalign.cc - I'm not sure that it is functionally correct,
but at least it compiles again.