The configuration file parser (and stereo.default files in general) has been accumulating old, no-longer used options for quite some time, and I finally took a crack today at cleaning things up.  Gone now are the last vestiges of the old C-style code for parsing those files, and everything has been consolidated into a singleton class called StereoSettings that you can access by calling the stereo_settings() method.  This method can be called from anywhere, so we no longer need to worry about passing the dft struct around the code.  Most importantly, I have removed many (close to 100!) deprecated options that no longer do anything.
The new stereo settings class will issue a warning for any deprecated
option it detects in the file.  For old files with many deprecated
options, there will be *many* such unknown options.  I've added a
stereo.default.example file to the top level of the Stereo Pipeline
repository with all of the current options (and some reasonable
default values for them).  Many of you can probably take this file,
change a few of the settings, and start using it.  If you want to use
your old files, be careful to peruse those warning.  (The change from
XCORR_TRESHOLD to XCORR_THRESHOLD is particularly subtle, so watch out
for that one!)

This much needed clean-up has been a long time in coming!  I hope it
makes things more clear!