revert putting the system path last
okay, so this is annoying. there's really no way to detect whether
we found something in the system path or not. We can't use something like
-nostdinc without making things really hard.

The only safe thing to do to use a non-standard path when a pkg is in your
standard path is to pass --with-pkg and be sure it's actually there.