Fixed problem where linescan imagers stopped working for isis_adjust
The problem was that pixel sigmas (which are in mm in isis_adjust)
were be turned to zero during unit conversion. Our previous code that
tried to guess by comparing mm measurements at the start and stop of an
image was returning zero (which they are in a linescan camera). Now we
refer to isis camera model for the pixel resolution. Though this
doesn't give quite the same pixel to mm ratio as before, I'm assuming it
is correct or good enough since it's from Isis.

Will add isis_adjust example back into documentation. Sorry Ross and Michael.