Turning on subpixel mode in example. Fixing bug in Holefilling(ish).
Problem was in detection. The bug looks like the inverse of a 'cut in'.
It's a square edge bump out in the disparity that is smooth. It looks like
a curled paper edge in the 3D model. I don't understand why I haven't noticed
it before, but hopefully it wasn't causing Kyle too much grief. It kinda
looks like a problem that would happen in the correlator.