Added custom projection into linescan to replace LineScanCameraGroundMap
I feel a bit awful about this but on all version of IsisCameraModel (
even the original by Michael ) was not able to correctly project a point
into a MOC-NA imager using the new unit tests. I since wrote my own
projection algorithm into a linescan image using the default LMA in VW.
LineScanCameraGroundMap fails to project the point because it throws a
Bounds Error, it means it thinks the point is outside the linescan imager.
I wasn't able to debug past that point and just gave up; that's when I wrote
my own algorithm.

The use of our own LMA has a positive note. We finally have the potential to use
IsisAdjustCameraModel with LineScan cameras without using the hack of storing
ephemeris time like we do inside of isis_adjust.

This commit also optimizes projection into frame cameras by not using
ground maps. This avoids a conversion to Lon Lat Radius and planet backside
checking.