Added IsisInterface skeletons
This is to be like DiskImageResource. IsisCameraModel should use the
IsisInterface files to perform it operations. The interface files provide
custom optimized code that goes out of its way to avoid invoking the entire
ISIS camera model. This is especially helpful for Map Projected files as the
SetImage command will do a ground intersection twice with every call! Eck.
Unfortunately these will all have to be heavily tested to make sure that ISIS
doesn't change too much to break this code.