Map Frame example up.
My previous check in was wrong .. perfomance boost still only 3-4x.
MapFrame still has issues. Point_to_pixel might be deviating.

For point to pixel, I probably need to first project into the camera
and then back down on to the map to take in account that the point
can have arbituary radius and not be fixed to the surface. Eck.