Added RemoteProjectFile and phoinitet
RemoteProjectFile allows for easy access to the project files that ptk_server is dish'n.

Phoinitet at this point is just a place holder. When we're not solving for the reflectance, phoinitet should immediately exit. For the time being, phoinitet will just print out debugging information of the requested ptk file. See the documentation to see where this file fits in the plan.