Added a ErodeView.
This is used to erase blobs that were detected with BlobIndexThreaded. This is applied because after stereo's filtering (which removes high gradients), there are left a whole bunch of little islands of valid data. This data is probably not correct and it's best to remove them. By default stereo deletes islands with an area of 1000 px.

At the moment ErodeView is a little slow. It needs an R-Tree or something similar to speed up it's seaching through blobs.