Added macros to catch VW and STD errors
isis_adjust and orbitviz have been switched over to using the new format. I hope to switch over the rest of the executables to this new format. An important point of this catch is that it catches a malloc failure so that it will tell the users directly and understandably that they are out of RAM.