Added Universal center option
There's a forgotten setting called universal radius that's used to
filter points that are too close or too far from the camera. Universal
center option allows the user to select the center of the filter. It
can be either CAMERA (which does what the code did before) or
ZERO. That last one is useful for filtering points that couldn't
realistically be on the surface of the planet.