Add asp/Core/Common.h
The purpose of Common.h is to standardize the input options and how
GDAL objects are written. This will hopefully make sure that BigTIFF
is on at all times.

Common provides the options: thread, help, and no-bigtiff. It also
provides simplified block writing. Finally it make sure there is an
error check for command line parsing.