Added a K10 example set
This is mostly just so that we can verify that we haven't broken
K10. I have a habit of forgeting about them. This is also a nice small
example that runs faster than MOC. All of our features also don't
currently work with K10. So here's hoping I'll remember to work on it
when free time comes around.