Added Threaded Edge Mask
This is a thread safe edge mask view ( i.e. it masks the black around
the sides of an image ). It starts a parallel process on construction
to find edges. Then on write, each block takes a small copy of the
global vectors to perform rasterization.

This is 40 second improvement for our MOC demo.

We don't just threshold on zero since we are attempting to preserve
black at the interior of the image. However, I don't understand why we
don't just use a nodata.