Clean up and fix point2dem
There was an overflow problem happening when calculating the mean
position of the point cloud. This is potentially the crash the users
were seeing with HiRISE input.

Also, I clean up the output of the executable. It looks more inline
with the other tools.

Also, DRG w/ alpha should work now.