Massive speed optimizations to OrthoRasterizer
With every block rasterization of OrthoRasterizer, it was iterating
over the entire point cloud! The ENTIRE POINT CLOUD! That was bad.

I've subdivided the input point cloud into bounding boxes to avoid
iterating over the entire input. This reduced the computations to
1/65536th of before.

Next best thing would be to implement a quad tree.