Core: Added active_area to ThreadedEdgeMask
Active area just provides a BBox that defines where the active area of
the image is. The search step also changes with image size now. This
is partly just to make my small image tests work, but it also means we
can keep processing time for large images to be somewhat linear. The
bigger the image .. the bigger the search step.