Core: IP Matching Tweaks (don't pick empty set cluster)
We now check that the variance is non-zero of another set that we
think has a lower variance. This because some of the DG images that we
have, seem to cause all inliers for the images!

There are some VW_OUT() changes here that also make the text output a
little cleaner.