core: Reduce IP clustering to 1 dimension
It seems the interest point code still found a way to over constrain
itself. Results have improved for one of my CTX examples if I instead
do the clustering of IP in one dimension. I achieve that take the
vector distance of a sample after have normalized and scaled all the
samples to be centered on zero and have an average distance of 1.2 in
all directions.