Core: Fix datum intersection for IP matching
This is a fix for Shean's RANSAC failure. The problem was that I
skipped on the math, I was intersecting a sphere instead of a
flattened one. That's a big mistake on Earth.

Anywho .. I've added tests to verify assumptions and to make it easier
if someone wants to play with that code.