DG: Use optimal threshold IP matching for map projected images
We're doing this already for nonmap projected images. This commit
allows it to happen for map projected images as well. The only thing
that needed to be shut off was the rough homography alignment. That
strategy doesn't apply well to map projected images and would require
a lot processing time.

Side note: This also added a RPCMapTransform object. This basically
everything you need in VW to actuall start writing a RPC Map
Projection tool.