stereo_mpi: Added a multiprocess variant of stereo
This leverages the same stereo executables as the python stereo script
... however it uses the left crop window option to divy up jobs among
multiple processes instead of threads. It then merges the results back
together using GDAL build vrt.

In the future this executable will hopefully have a 'local' mode and a
'mpiexec' mode for the super computer. This will be our answer for
processing a stereo pair across multiple computers.