stereo: Added stereo_corr_sub
This separates out D_sub calculation so that it can optionally happen
in stereo_corr_sub. This is important for MPI calculation so that
D_sub is not generated many processes over. This might also be the
point in the code where we make D_sub from DEM generation possible.