Core: Fix mistake introduced in bdf4a7b
I removed one too many allocations! This copy of the input image is
actually very important as we were dealing with a lazy view up to this
point. Integral Image operation is a per pixel heavy operation that
access each pixel twice. It was causing havok on the cache system
prior to this commit and after bdf4a7b.