pprc: Have sub images rendered in parallel with sub_tile_size
Sub_tile_size and sub_theads are still important to make sure that
rendering the low res images doesn't allocate a ton of
GB. Cache_tile_aware_render stop us from notice we had a
problem. However if the sub_scale ration is greater than five ... or
if there just existed a really weird homography, it was possible that
stereo_pprc would try to allocate many GB of ram. This brings down the
system because it is swapping memory so hard.