session: Add DG Map RPC Session
This is a session specifically made so that we don't use a
TransformRef. This allows multiple copies of RPCMapTransform to be
created for each thread (w/ TransformRef a single copy would be hidden
behind a shared ptr). These multiple copies thus allow RPCMapTransform
to cache its DEM usage ... and thus Triangulation happens faster.