RPC: Revert 04f0645
Returning negative numbers is a bad idea as it causes raster bounding
boxes to always include that pixel. This makes raster buffer inside
TransformView really large and was causing David's memory problems.

Oleg didn't catch on to this because when he wrote that code he was
using RPCMapTransform with the Convex option turned on. (Something I
did).