sparse_disp: move gridding to its own routine, allow an optional argument to specify the dilation distance for the valid mask
Gridding the output disparities was pretty ad hoc.  Now there's a formal process to ensure that disparities and spreads don't get interpolated too far, and to check the distance between the grid points and the data points.  An optional argument can specify the dilation of the mask around the valid data points.