sparse_disp: Add maximum-disparity-cluster test
This option improves on the maximum-disparity-distance test in previous versions.  Now we identify all matches with C> C_tol and establish how well they are clustered in disparity space based on their minimum disparity difference with a neighbor.  The 90th percentile of the clustering distribution in each direction is used as a filter on the matches with C< C_tol.  This tends to eliminate spatailly isolated blunders with poor correlation score.  This test is enabled by default, disabled by invoking the --Exhaustive option.