sparse_disp: local_epipolar option, cleanup
For very large images, the distortions introduced by rpc mapping can lead to variations in the effective epipolar axis.  This commit breaks the image into large chunks and estimates an epipolar vector for each, falling back to the global epipolar vector if there are not enough high-quality points available.

This commit also removes the ep_vec_initial option, which was mostly unused, and cleans up some sloppy filtering statements in the code.