Use Ceres's DynamicCostFunction in bundle_adjust.cc
Previously we were at compile time instantiating every possible number of
parameters for camera intrinsics for the reprojection cost function. Ceres
provides another way were this can be done at run time using the
DynamicNumericCostFunction. This should just be a drop in replacement with no
change in functionality. This makes the code a little smaller and hopefully a
little bit easier to read.