Fix clang warnings in ASP.
We had one unused variable and several structs accidentally forward
declared as classes.

In BaseEquation(), clang was confused and saw an infinite loop in
operator[]. That was fixed with a const_cast.

Finally, don't pass POD types with a const&. POD types are the same size
as a pointer so it's faster to just copy the POD type. See the first
comment of:
https://stackoverflow.com/questions/7561690/why-do-we-not-pass-pod-by-reference-in-functions

The internal C++ guide for Google also complains about this.
Unfortunately their public style guide doesn't mention it. :/